Refactor the code into a single node project, that will build the client and server all together.
