+++ .gitignore
@@ -1,3 +1,4 @@
 node_modules/
 .idea/workspace.xml
 /gateway/sync_db
+yarn.lock
