--- gateway/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "experiment",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "build": "./node_modules/.bin/babel  ./src -d ./dist",
-    "start": "BABEL_ENV=development node_modules/.bin/nodemon src/main.js --exec node_modules/.bin/babel-node --presets env,es2015"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "nodemon": "^1.17.1"
-  },
-  "dependencies": {
-    "body-parser": "^1.18.2",
-    "express": "^4.16.2",
-    "http": "0.0.0",
-    "nano": "^6.4.3",
-    "pouchdb": "^6.4.3",
-    "redux": "^3.7.2",
-    "serialport": "^6.1.1",
-    "url": "^0.11.0",
-    "usb-detection": "^2.1.0",
-    "ws": "^4.1.0"
-  }
-}
