--- kiosk/README.md
@@ -1,1 +1,0 @@
-npm install && npm start
