+++ scripts/start.js
