+++ src/client/lib/redux-share-client.js
