+++ src/client/pages/index.js
