+++ src/client/pages/logo.png
