+++ src/client/pages/splash.js
