+++ src/client/public/favicon.ico
