+++ src/server/barcode/BarcodeReader.js
